+++ CHANGELOG.md
@@ -3,6 +3,7 @@ v1.5.1 [unreleased]
 
 ### Bugfixes
 -   [#204](https://github.com/latera/homs/pull/204) Pass locale for translations from environment.
+-   [#211](https://github.com/latera/homs/pull/211) Fix lookup and change multi select to single.
 
 v1.5.0 [2018-10-29]
 -------------------